requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.66.1
PyYAML==6.0.1
tensorboard==2.14.0
opencv-python==4.8.0.76
albumentations==1.3.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pytest==7.4.2
Flask==2.3.2
